Moved things around, there will be more refactoring comming soon.

+++ templates/remake/build.sh
@@ -8,6 +8,5 @@ gcc -O2 -march=x86-64-v2 -mtune=native -shared -fPIC -o "$OUT_FILE".so remake.c 
 # Build Windows version
 x86_64-w64-mingw32-gcc -O2 -march=x86-64-v2 -mtune=native -shared -o "$OUT_FILE".dll remake.c -I../../include
 
-mkdir -p ../../bin/remakes
-mv "$OUT_FILE".so ../../bin/remakes
-mv "$OUT_FILE".dll ../../bin/remakes
+[ -e "$OUT_FILE".so ] && mv "$OUT_FILE".so ../../bin/remakes
+[ -e "$OUT_FILE".dll ] && mv "$OUT_FILE".dll ../../bin/remakes
